requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
grpcio==1.18.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
